+++ src/earthengine/index.js
@@ -1,0 +1,3 @@
+import getEarthEngineWorker from './ee_worker_loader'
+
+export default getEarthEngineWorker
